[refactor] 테스트 클래스 위치 수정
