+++ src/test/java/racingcar/CarTest.java
@@ -29,11 +29,5 @@ class CarTest {
         assertThat(car.getDistance()).isEqualTo(0);
     }
 
-    @Test
-    @DisplayName("빈 자동차에서 최대 거리를 구하려는 경우 예외를 발생시킨다.")
-    public void emptyCars() {
-        Cars cars = new Cars(new String[]{});
-        assertThatThrownBy(cars::getMaxDistance)
-                .isInstanceOf(RuntimeException.class);
-    }
+
 }
